servicios del back nuevos agregados

+++ backend/src/main/java/com/velonet/backend/dto/CurvaPagoMensualDTO.java
@@ -1,13 +1,11 @@
 package com.velonet.backend.dto;
 
-import java.math.BigDecimal;
-
 //DTO para enviar al front para la curva
 public class CurvaPagoMensualDTO {
     private String mesAnio; // Formato: "yyyy-mm"
-    private BigDecimal totalCobrado;
+    private Double totalCobrado;
 
-    public CurvaPagoMensualDTO(String mesAnio, BigDecimal totalCobrado) {
+    public CurvaPagoMensualDTO(String mesAnio, Double totalCobrado) {
         this.mesAnio = mesAnio;
         this.totalCobrado = totalCobrado;
     }
@@ -15,6 +13,6 @@ public class CurvaPagoMensualDTO {
     //Getters y Setters
     public String getMesAnio() { return mesAnio; }      
     public void setMesAnio(String mesAnio) { this.mesAnio = mesAnio; }
-    public BigDecimal getTotalCobrado() { return totalCobrado; }
-    public void setTotalCobrado(BigDecimal totalCobrado) { this.totalCobrado = totalCobrado; }
+    public Double getTotalCobrado() { return totalCobrado; }
+    public void setTotalCobrado(Double totalCobrado) { this.totalCobrado = totalCobrado; }
 }
